# dist/openapi-complete.yaml
openapi: 3.0.0
info:
  title: המתכונים של סבתא ואחרים - Grandma's Recipes and Others API
  description: Complete API for recipes project including authentication, search, personal area, and bonus features
  version: 2.0.0
servers:
  - url: "http://127.0.0.1:80"
    description: Local development server
...
paths:
...
  /users/{userId}/favorites:
    get:
      tags:
        - Favorites
      summary: קבלת מתכונים מועדפים
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's favorite recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipePreview"
      security:
        - cookieAuth: []
    
    post:
      tags:
        - Favorites
      summary: הוספת מתכון למועדפים
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
      responses:
        "200":
          description: Recipe added to favorites
      security:
        - cookieAuth: []
  /users/{userId}/favorites/{recipeId}:
    delete:
      tags:
        - Favorites
      summary: הסרת מתכון מהמועדפים
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe removed from favorites
      security:
        - cookieAuth: []
...
  /users/{userId}/recipes:
    get:
      tags:
        - Personal
      summary: קבלת מתכונים אישיים
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's personal recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRecipe"
      security:
        - cookieAuth: []

    post:
      tags:
        - Personal
      summary: יצירת מתכון אישי חדש
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipe"
      responses:
        "201":
          description: Recipe created successfully
      security:
        - cookieAuth: []
  /users/{userId}/recipes/{recipeId}:
    delete:
      tags:
        - Personal
      summary: מחיקת מתכון אישי
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe deleted successfully
      security:
        - cookieAuth: []
  /users/{userId}/family-recipes:
    get:
      tags:
        - Family
      summary: קבלת מתכונים משפחתיים
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's family recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyRecipe"
      security:
        - cookieAuth: []

    post:
      tags:
        - Family
      summary: יצירת מתכון משפחתי חדש
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyRecipe"
      responses:
        "201":
          description: Family recipe created successfully
      security:
        - cookieAuth: []
  /users/{userId}/search/last:
    get:
      tags:
        - Search
      summary: קבלת החיפוש האחרון של המשתמש
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Last search query
      security:
        - cookieAuth: []
  /users/{userId}/recipes/{id}/progress:
    get:
      tags:
        - Bonus - Preparation
      summary: קבלת התקדמות הכנת מתכון
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe preparation progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeProgress"
      security:
        - cookieAuth: []

    put:
      tags:
        - Bonus - Preparation
      summary: עדכון התקדמות הכנת מתכון
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                current_step:
                  type: integer
                is_completed:
                  type: boolean
                notes:
                  type: string
      responses:
        "200":
          description: Progress updated successfully
      security:
        - cookieAuth: []
  /users/{userId}/meal-plan:
    get:
      tags:
        - Bonus - Meal Planning
      summary: קבלת תכנון ארוחות
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Meal plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlanItem"
      security:
        - cookieAuth: []

    post:
      tags:
        - Bonus - Meal Planning
      summary: הוספת פריט לתכנון ארוחות
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealPlanItem"
      responses:
        "201":
          description: Meal plan item added
      security:
        - cookieAuth: []
  /users/{userId}/meal-plan/{itemId}:
    put:
      tags:
        - Bonus - Meal Planning
      summary: עדכון פריט בתכנון ארוחות
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMealPlanItem"
      responses:
        "200":
          description: Meal plan item updated
      security:
        - cookieAuth: []

    delete:
      tags:
        - Bonus - Meal Planning
      summary: מחיקת פריט מתכנון ארוחות
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meal plan item deleted
      security:
        - cookieAuth: []
components:
  schemas:
...
    RecipePreview:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 102
        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        readyInMinutes:
          type: integer
          description: time to make this recipe
          example: 45
        aggregateLikes:
          type: integer
          description: how many likes this recipe got
          example: 5987
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false
        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
...
    UserRecipe:
      type: object
      properties:
        recipe_id:
          type: integer
        title:
          type: string
        image:
          type: string
        readyInMinutes:
          type: integer
        servings:
          type: integer
        instructions:
          type: string
        ingredients:
          type: string
        is_vegetarian:
          type: boolean
        is_vegan:
          type: boolean
        is_gluten_free:
          type: boolean
        created_at:
          type: string
          format: date-time
    FamilyRecipe:
      allOf:
        - $ref: "#/components/schemas/UserRecipe"
        - type: object
          properties:
            family_origin:
              type: string
            family_member:
              type: string
            when_to_prepare:
              type: string
    CreateRecipe:
      type: object
      required:
        - title
        - instructions
        - ingredients
      properties:
        title:
          type: string
        image:
          type: string
        readyInMinutes:
          type: integer
        servings:
          type: integer
        instructions:
          type: array
          items:
            type: string
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: number
              unit:
                type: string
        is_vegetarian:
          type: boolean
        is_vegan:
          type: boolean
        is_gluten_free:
          type: boolean
    CreateFamilyRecipe:
      allOf:
        - $ref: "#/components/schemas/CreateRecipe"
        - type: object
          properties:
            family_origin:
              type: string
            family_member:
              type: string
            when_to_prepare:
              type: string
    RecipeProgress:
      type: object
      properties:
        recipe_id:
          type: integer
        current_step:
          type: integer
        is_completed:
          type: boolean
        notes:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    MealPlanItem:
      type: object
      properties:
        item_id:
          type: integer
        recipe_id:
          type: integer
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        planned_date:
          type: string
          format: date
        servings:
          type: integer
        order_index:
          type: integer
        created_at:
          type: string
          format: date-time
    CreateMealPlanItem:
      type: object
      required:
        - recipe_id
        - meal_type
        - planned_date
      properties:
        recipe_id:
          type: integer
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        planned_date:
          type: string
          format: date
        servings:
          type: integer
          default: 1
        order_index:
          type: integer
    UpdateMealPlanItem:
      type: object
      properties:
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        planned_date:
          type: string
          format: date
        servings:
          type: integer
        order_index:
          type: integer
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session